[clear] デザイン用のextentionを分離
