+++ SyncPod/ViewController/SignInViewController.swift
@@ -8,15 +8,6 @@
 
 import UIKit
 
-extension UITextField {
-    func addBorderBottom(height: CGFloat, color: UIColor) {
-        let border = CALayer()
-        border.frame = CGRect(x: 0, y: self.frame.height - height, width: self.frame.width, height: height)
-        border.backgroundColor = color.cgColor
-        self.layer.addSublayer(border)
-    }
-}
-
 class SigninViewController: UIViewController, UINavigationBarDelegate {
     @IBOutlet weak var navigationBar: UINavigationBar!
     @IBOutlet weak var mailField: UITextField!
